requirements.txt
crewai
crewai_tools
langchain
langchain-ollama
python-dotenv
litellm
PyPDF2
python-docx
chromadb
pydantic